[proofs] [alethe] Generalize ALPHA_EQUIV translation (#11236)
In some cases the renaming used in the ALPHA_EQUIV rule in cvc5 is not
directly valid via a `bind` rule in Alethe, since entries added to the
context are used to generate a substitution using a cumulative strategy.
This PR addresses this issue by breaking down the renaming into more
than one `bind` step when this happens. The translation in the Alethe
post-processor is generalized to account for both the regular and this
special case.