Add infrastructure to translate THEORY_LEMMAS that have a ProofRewriteRule + two distinct rules (#11399)
There are some rewrite rules that cannot be expressed in RARE (also
called ad-hoc rules). They should be translated to existing Alethe rules
whenever possible.

This PR adds a separate function to the callback for this which is
cleaner then doing it in the main update function because of too much
indentation.

It also translates the DISTINCT_ELIM and DISTINCT_CARD_CONFLICT rules.